+++ src/radarpub.rs
@@ -249,7 +249,12 @@ async fn stream(
 
                 let span = info_span!("targets_publish");
                 async {
-                    match targets_publisher.put(msg).encoding(enc).await {
+                    match targets_publisher
+                        .put(msg)
+                        .encoding(enc)
+                        .timestamp(session.new_timestamp())
+                        .await
+                    {
                         Ok(_) => {}
                         Err(e) => error!("{} publish error: {:?}", args.targets_topic, e),
                     }
@@ -420,7 +425,12 @@ async fn clustering_task(
 
         let span = info_span!("clusters_publish");
         async {
-            match publisher.put(msg).encoding(enc).await {
+            match publisher
+                .put(msg)
+                .encoding(enc)
+                .timestamp(session.new_timestamp())
+                .await
+            {
                 Ok(_) => {}
                 Err(e) => error!("{} message error: {:?}", args.clusters_topic, e),
             }
@@ -599,7 +609,12 @@ async fn cube_loop(
                         let (msg, enc) = format_cube(cubemsg, &frame_id).unwrap();
                         let span = info_span!("cube_publish");
                         async {
-                            match cube_publisher.put(msg).encoding(enc).await {
+                            match cube_publisher
+                                .put(msg)
+                                .encoding(enc)
+                                .timestamp(session.new_timestamp())
+                                .await
+                            {
                                 Ok(_) => {}
                                 Err(e) => error!("publish cube error: {:?}", e),
                             }
@@ -707,15 +722,21 @@ async fn tf_static(
     msg: ZBytes,
     enc: Encoding,
 ) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-    let topic = "rt/tf_static".to_string();
+    let topic = "tf_static".to_string();
     let mut interval = tokio::time::interval(Duration::from_secs(1));
 
     loop {
         interval.tick().await;
         let span = info_span!("tf_static_publish");
-        async { session.put(&topic, msg.clone()).encoding(enc.clone()).await }
-            .instrument(span)
-            .await?;
+        async {
+            session
+                .put(&topic, msg.clone())
+                .encoding(enc.clone())
+                .timestamp(session.new_timestamp())
+                .await
+        }
+        .instrument(span)
+        .await?;
     }
 }
 
@@ -724,15 +745,21 @@ async fn radar_info(
     msg: ZBytes,
     enc: Encoding,
 ) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-    let topic = "rt/radar/info".to_string();
+    let topic = "radar/info".to_string();
     let mut interval = tokio::time::interval(Duration::from_secs(1));
 
     loop {
         interval.tick().await;
         let span = info_span!("radar_info_publish");
-        async { session.put(&topic, msg.clone()).encoding(enc.clone()).await }
-            .instrument(span)
-            .await?;
+        async {
+            session
+                .put(&topic, msg.clone())
+                .encoding(enc.clone())
+                .timestamp(session.new_timestamp())
+                .await
+        }
+        .instrument(span)
+        .await?;
     }
 }